+++ client/src/pages/Main/Main.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import MainNav from "@/components/main/MainNav/MainNav";
 import PostScroll from "@/components/main/PostScroll/PostScroll";
 import useModalStore from "@/store/useModalStore";
-import TagRank from "@/components/main/TagRank/TagRank";
+import TagRanking from "@/components/main/TagRanking/TagRanking";
 
 import "./Main.scss";
 
@@ -17,7 +17,7 @@ const Main = (): JSX.Element => {
       <MainNav />
       <div className="main__content">
         <PostScroll />
-        <TagRank />
+        <TagRanking />
       </div>
     </div>
   );
